llvm: Use host-native llvm-tblgen when building for target
During LLVM builds, the llvm-tblgen program needs to be available. This
can be solved by making "llvm" depend on "host-llvm", and use the CMake
LLVM_TABLEGEN build setting to the version built as part of "host-llvm".

It is a bit of overkill to build the whole of "host-llvm" only to get a
working llvm-tblgen, but it gets the job done with minimum fuss. At some
point it would be desirable to only build that tool unless some other
package has a need for a full "host-llvm" build.