llvm: Build a llvm-config for the host which provides info about the target
Add commands to reconfigure the subdirectory where LLVM builds native
tools, fixing them up to use the host-native compiler instead of the
Buildroot cross-toolchain. This way the built "llvm-config" program can
run in the host, and provide information about the target build of LLVM
(which is the one available in the sysroot).