Add wpebackend-shm and allow selecting multiple backends
This adds the shm WPE backend. In order to support building multiple
backends, and being able to choose among them at runtime, this also
adds:

- A (hidden) virtual package "wpebackend-implementation" for other
  packages to depend upon.
- The possibility to build multiple WPE backend implementations.
- A choice of default backend, which also controls the
  "WPEBackend-default.so" symlink.

The last bit (symlink fixup) is done in a separate hidden package which
only installs the symlink in the target, and depends on all the enabled
backends in order to ensure the commands are executed after all have
been installed. This is needed because some backends add a symlink
themselves when built, and the package-installed symlinks must be
overwritten.