[opencdmi] Avoid ambiguous revision
The git revision specified before this patch exists in several branches,
so it's not clear at all which branch even gets built...

I have switched to a branch build. This means new commits will go live
so-to-speak automatically, but that should not be an issue since this is
a reference branch.